Migration to Gatling 3.0.0, simplify source code layout.
