--- sample/gatling-layout/build.gradle
@@ -1,7 +1,0 @@
-plugins {
-    id 'com.github.lkishalmi.gatling' version '0.7.1'
-}
-
-repositories {
-    jcenter()
-}
