+++ src/test/groovy/functional/WhenCompileSimulationSpec.groovy
@@ -1,6 +1,6 @@
 package functional
 
-import groovy.io.FileType
+import helper.GatlingFuncSpec
 import org.gradle.testkit.runner.BuildResult
 import org.gradle.testkit.runner.UnexpectedBuildFailure
 import spock.lang.Unroll
@@ -11,106 +11,57 @@ class WhenCompileSimulationSpec extends GatlingFuncSpec {
 
     static def GATLING_CLASSES_TASK_NAME = "gatlingClasses"
 
-    @Unroll
-    def "should compile, layout `#layout`"() {
+    def "should compile"() {
         setup:
-        prepareTest(layout)
-
+        prepareTest()
         when:
         BuildResult result = executeGradle(GATLING_CLASSES_TASK_NAME)
-
         then: "compiled successfully"
         result.task(":$GATLING_CLASSES_TASK_NAME").outcome == SUCCESS
-
         and: "only layout specific simulations were compiled"
         def classesDir = new File(testProjectBuildDir, "classes/scala/gatling")
         classesDir.exists()
-        and:
-        classesDir.eachFileRecurse(FileType.FILES) { assert it.name.contains(simulationPart) }
-
         and: "only layout specific resources are copied"
         def resourcesDir = new File(testProjectBuildDir, "resources/gatling")
         resourcesDir.exists()
-        new File(resourcesDir, resourceFile).exists()
-
+        new File(resourcesDir, "search.csv").exists()
         and: "main classes are compiled"
         def mainDir = new File(testProjectBuildDir, "classes/java/main")
         mainDir.exists()
-
         and: "test classes are compiled"
         def testDir = new File(testProjectBuildDir, "classes/java/test")
         testDir.exists()
-
-        where:
-        layout   || simulationPart || resourceFile
-        "gradle"  | "1Simulation"   | "search1.csv"
-        "gatling" | "2Simulation"   | "search2.csv"
     }
 
     @Unroll
-    def "should not compile without #dir in #layout layout"() {
+    def "should not compile without #dir"() {
         setup:
-        prepareTest(layout)
-        and: "remove test classes"
+        prepareTest()
+        and:
         assert new File(testProjectDir.root, dir).deleteDir()
-
         when:
         executeGradle(GATLING_CLASSES_TASK_NAME)
-
         then:
         thrown(UnexpectedBuildFailure)
-
         where:
-        layout    | dir
-        'gradle'  | 'src/main'
-        'gradle'  | 'src/test'
-        'gatling' | 'src/main'
-        'gatling' | 'src/test'
+        layout   | dir
+        'gradle' | 'src/main'
+        'gradle' | 'src/test'
     }
 
-    @Unroll
     def "should not compile without gatling dependencies"() {
         given: "build script without gatling dependencies"
-        prepareTest(layout).text = """plugins {
+        new File(testProjectDir.root, "build.gradle").text = """
+plugins {
     id 'com.github.lkishalmi.gatling'
 }
 repositories {
     jcenter()
 }
 """
-
         when:
         executeGradle(GATLING_CLASSES_TASK_NAME)
-
         then:
         thrown(UnexpectedBuildFailure)
-
-        where:
-        layout << [ 'gradle', 'gatling']
-    }
-
-    @Unroll
-    def "should compile simulations without `bodies` folder, layout `#layout`"() {
-        setup:
-        prepareTest(layout)
-        and: "remove `bodies` folder"
-        assert new File(testProjectDir.root, "src/gatling/$bodiesFolder").deleteDir()
-
-        when:
-        BuildResult result = executeGradle(GATLING_CLASSES_TASK_NAME)
-
-        then:
-        result.task(":$GATLING_CLASSES_TASK_NAME").outcome == SUCCESS
-
-        and: "only layout specific simulations were compiled"
-        def classesDir = new File(testProjectBuildDir, "classes/scala/gatling")
-        classesDir.exists()
-        and:
-        classesDir.eachFileRecurse(FileType.FILES) { assert it.name.contains(simulationPart) }
-
-        where:
-        layout    | bodiesFolder                   | simulationPart
-        'gradle'  | 'resources/bodies' | '1Simulation'
-        'gatling' | 'bodies'           | '2Simulation'
     }
 }
