--- src/test/resources/gatling-layout/src/gatling/data/search2.csv
@@ -1,3 +1,0 @@
-searchCriterion,searchComputerName
-Macbook,MacBook Pro
-eee,ASUS Eee PC 1005PE
