--- src/test/resources/gatling-layout/src/main/java/computerdatabase/MainUtils.java
@@ -1,7 +1,0 @@
-package computerdatabase;
-
-public class MainUtils {
-    public static String hostName() {
-        return System.getProperty("gatling.hostName");
-    }
-}
