--- src/test/resources/gatling-layout/src/test/java/computerdatabase/TestUtils.java
@@ -1,7 +1,0 @@
-package computerdatabase;
-
-public class TestUtils {
-    public static String hostName() {
-        return MainUtils.hostName();
-    }
-}
